+++ src/main/kotlin/com/github/surpsg/diffcoverage/services/notifications/BalloonNotificationService.kt
@@ -2,11 +2,7 @@ package com.github.surpsg.diffcoverage.services.notifications
 
 import com.github.surpsg.diffcoverage.DiffCoverageBundle
 import com.github.surpsg.diffcoverage.properties.PLUGIN_NAME
-import com.intellij.notification.NotificationDisplayType
-import com.intellij.notification.NotificationGroup
-import com.intellij.notification.NotificationListener
-import com.intellij.notification.NotificationType
-import com.intellij.notification.Notifications
+import com.intellij.notification.*
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.project.Project
 
@@ -14,13 +10,12 @@ import com.intellij.openapi.project.Project
 class BalloonNotificationService(private val project: Project) {
 
     fun notify(
-        group: NotificationGroup = DIFF_COVERAGE_BALLOON,
         notificationType: NotificationType = NotificationType.INFORMATION,
         notificationListener: NotificationListener? = null,
         message: String
     ) {
         Notifications.Bus.notify(
-            group.createNotification(
+            getNotificationGroup().createNotification(
                 DiffCoverageBundle.message(PLUGIN_NAME),
                 message,
                 notificationType,
@@ -30,11 +25,7 @@ class BalloonNotificationService(private val project: Project) {
         )
     }
 
-    companion object {
-        val DIFF_COVERAGE_BALLOON = NotificationGroup(
-            "diff.coverage.notification.balloon",
-            NotificationDisplayType.BALLOON,
-            true
-        )
+    private fun getNotificationGroup(): NotificationGroup {
+        return NotificationGroupManager.getInstance().getNotificationGroup("diff.coverage.notification")
     }
 }