+++ CHANGELOG.md
@@ -3,5 +3,11 @@
 # diff-coverage-idea-plugin Changelog
 
 ## [Unreleased]
+### Fixed
+- Intellij 2021.1 compatibility
+### Changed
+- Code coverage classes list flat structure
+
+## [0.0.3]
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)